Modifications to suppress sign-compare warnings

+++ src/System/DataLogger.cpp
@@ -130,7 +130,7 @@ bool DataLogger::open(String name, int numColumns) {
             int fileCountStart = base.lastIndexOf('(');
             int fileCountEnd = base.lastIndexOf(')');
             int count = 0;
-            if (fileCountStart != -1 && fileCountEnd != -1 && fileCountEnd == base.length() - 1) {
+            if (fileCountStart != -1 && fileCountEnd != -1 && fileCountEnd == (int)base.length() - 1) {
                 String countString = base.substring(fileCountStart + 1, fileCountEnd);
                 bool numeric = true;
